Applying Seperation of Concerns principle
The refactoring method used here follows modularization and separation of concerns. The idea is to break down the responsibilities of map.jsx into smaller, reusable components. This enhances readability, maintainability, and efficiency.
